Window x64 editor #179

+++ engine/core/util/object_pool.h
@@ -34,7 +34,7 @@ namespace Echo
 
 		void deleteObj(T* obj)
 		{
-			i32 idx = obj - m_objs;
+			i32 idx = static_cast<i32>(obj - m_objs);
 			m_frees.push_back(idx);
 		}
